Rename DOS contest page submodules.
