+++ client/src/component/DOS/Contest/DetailPage.tsx
